TestSessionLabelProviderCustom shows JUnit 5 TestTemplate poorly
When org.eclipse.jdt.junit.runners is included in a product, JUnit 5
TestTemplate tests are not displayed correctly in the JUnit view. E.g. a
RepeatTest is displayed as follows:

test.TestTemplatesJunit5
* test.TestTemplatesJunit5
** test1

Without this plug-in, the JUnit view has more useful contents:

TestTemplatesJunit5
* test1()
** repetition 1 of 1

This is due to outdated contents in TestSessionLabelProviderCustom, when
compared to TestSessionLabelProvider. Only the latter was updated to
deal with JUnit 5 tests.

This change adjusts TestSessionLabelProviderCustom.getSimpleLabel() to
the change in TestSessionLabelProvider.getSimpleLabel() for:
https://bugs.eclipse.org/bugs/show_bug.cgi?id=488566

Fixes: #18